Setup event loop for packet relayer
